move login route from server to login.js
